requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0